v2 Delta 1.0 || Initialted a new Page [Modal Page]
The [Modal Page] will contain a simple primary button asking the user to show a Modal. If the button is clicked,
then the background will get little dark and a modal box will pop up showing a message passed to the [Modal] component.
This modal box will also have a button to acknowledge the message and closing the modal. The modal box will also close if
use clicks anywhere outsite the modal box.