v2 Delta 1.5 || Created Modal Background & its logics
Created a a div in the Modal component that accuires the entire screen and gives a dark background.
That div has a click handler assigned to it, that changes a piece of state, and invokes a useEffect function,
 that checks if the click is done outside the modal box or not. If the click is inside the modal box the useEffect function
sets the two click hander state to false again. If the click is done outsite the modal box, the useEffect function
 invokes a function [closeModal()] that is being passed to Modal component as a prop.