v3 Delta 1.6 || Made the rows dynamic
Previously the 'td' elements were hard coded and was fetching all the values from the dataset.
Now so that we have the config data, and a variable column, the rows should adapt it too. Hence
this commit makes the rows responsive to the number of column we have on config file.
This code simply maps over the [data] array and individual objects are being called [row], inside that
mapped function we have another mapping of the [condig] array and each objects being called [colConf],
now inside that mapped function we have the 'td' element with a assign to the render function of the
colConf object. Then in the return function of the first map function we have the 'tr' element and inside
that element we have the [renderCells] const called, which containes the second mapping.