v3 Delta 2.4 || Created mechanism for handling header icons
Created some mechanism for handling the icons in the table headers. If the column is sortable, then a up-down arrow is rendered
then if the column is string then a icon with A and Z rendered on sorting, and 1,9 for number types.
When sorting ascending, the arrow incades A to Z or 1 to 9 and vice versa.