v3 Delta fix 2.4.2 || Fixed sortOrder cycle
Previously the sortOrder cycle was not breakable and was cycling on the same manner, without depending on anything else
. Which was creating a big problem. Whenever user clicked on a sortable header and sortOrder became 'asc' or 'desc' and then
if user clicked on any other sortable header, that column's sortOrder was following after that previous one's sortOrder.
So, the sortOrder was undependent of the column that is being sorted. In this commit, that changes created a mechanism, which
allows the the sortOrder to look after, which column is being sorted. If user clicks on any other column than that is being sorted,
the sortOrder restarts it's cycle.