v3 Delta 2.6 || Better design for determining Column's Data type
Previously we used a hardcoded standalone property in the config array for declaring the data type of a certain sortable
column. That was working fine, but was a extra headache for data entry/data updation. Hence this commit determines the data type
of a column by looking it's first element using the sortValue property of config data. Then this data type determation is used
to show appropriate icon on sorting. The only exception to this feature is that, it kind of expects to one column be only one data typew.
If the first element in a column is x data type and rest are y, the App will accept this column as a x data type column.