v3 Delta 2.9 || Fixed click avoidation in NavBar
Previously in the Link component that is used by the Navbar to show the navigation
 buttons was showing a anchor tag inside a div, styled to look like a button.
This was causing the button to be getting hovered but not invoking clicks if
the click is placed outside the button text. To resolve that issue, a button
element inside the anchor tag is added in this commit and styles of the parent
div is used, with a removation of the parent div, as the anchor tag itself can
be a parent tag.