v4 Delta 1.2 || Used useReducer in useCounter hook
The useCounter hook contains two states, which are similar in nature and used in similar
stuff. Hence restructured them using useReducer hook. And used the dispatch function to update
the reducer states by passing a object, having one property as [type] to declare type and another
as [payload] for the value. Then created one state setter function for [incrementBy] state and passed it
to children components by returning it in the returning object