v4 Delta 1.3 || Used Immer lib in useCounter custom hook
Used the Immer lib in our custom hook useCounter to reduce code.
This library allows to directly assing values to reducer states.
After assignation of values to a state, immer takes those values and
creates a obejct with all the other states and passes that updated
state. This object then gets passed with the dispatched function.