Version 4 Anaros 🍍 Release
The version 4 of the Adprava app has a new page , which is a basic counter component, famous in ReactJS World. The aim of this page is learning and practicing Reducer functions in ReactJS. In addition to that this page is also a practice for custom hooks, as all the logics behind the counter component is migrated to a reusable custom hook.